Add forgotten files and remove unused wildcard class
